Added scripts and patches to run CramSim with CPU models (GEM5, MacSim, Ariel) Updated the README file. Changed the data type of the time variables from signed (32-bit) to SimTime_t (64-bit)

+++ src/sst/elements/CramSim/c_BankCommand.cpp
@@ -91,7 +91,7 @@ c_BankCommand::c_BankCommand(unsigned x_cmdSeqNum,
 	m_cmdToString[e_BankCommandType::REF] = "REF";
 }
 
-unsigned c_BankCommand::getAddress() const {
+ulong c_BankCommand::getAddress() const {
 	return (m_addr);
 }
